Use generic list of floats in GraphView and move the class to other package
