+++ app/src/main/java/org/wikipedia/views/GraphView.kt
@@ -1,4 +1,4 @@
-package org.wikipedia.feed.view
+package org.wikipedia.views
 
 import android.content.Context
 import android.graphics.*
@@ -11,9 +11,9 @@ import org.wikipedia.feed.mostread.MostReadArticles
 class GraphView(context: Context, attributeSet: AttributeSet) : View(context, attributeSet) {
     private val gradientColor1 = ContextCompat.getColor(context, R.color.accent50)
     private val gradientColor2 = ContextCompat.getColor(context, R.color.green50)
-    private val dataSet = mutableListOf<MostReadArticles.ViewHistory>()
-    private var maxX = 0
-    private var maxY = 0
+    private val dataSet = mutableListOf<Float>()
+    private var maxX = 0f
+    private var maxY = 0f
     private val path = Path()
 
     private val pathPaint = Paint().apply {
@@ -24,11 +24,11 @@ class GraphView(context: Context, attributeSet: AttributeSet) : View(context, at
         isAntiAlias = true
     }
 
-    fun setData(viewHistoryList: List<MostReadArticles.ViewHistory>) {
-        maxX = viewHistoryList.size
-        maxY = viewHistoryList.maxByOrNull { it.views }?.views ?: 0
+    fun setData(list: List<Float>) {
+        maxX = list.size.toFloat()
+        maxY = list.maxByOrNull { it } ?: 0f
         dataSet.clear()
-        dataSet.addAll(viewHistoryList)
+        dataSet.addAll(list)
         invalidate()
     }
 
@@ -37,9 +37,9 @@ class GraphView(context: Context, attributeSet: AttributeSet) : View(context, at
         path.reset()
         dataSet.forEachIndexed { index, data ->
             if (index == 0) {
-                path.moveTo(index.scaleX() + GRAPH_MARGIN, data.views.scaleY() + (GRAPH_MARGIN / 2))
+                path.moveTo(index.scaleX() + GRAPH_MARGIN, data.scaleY() + (GRAPH_MARGIN / 2))
             } else if (index < dataSet.size) {
-                path.lineTo(index.scaleX() + GRAPH_MARGIN, data.views.scaleY() + (GRAPH_MARGIN / 2))
+                path.lineTo(index.scaleX() + GRAPH_MARGIN, data.scaleY() + (GRAPH_MARGIN / 2))
             }
         }
         canvas.drawPath(path, pathPaint)
@@ -52,7 +52,7 @@ class GraphView(context: Context, attributeSet: AttributeSet) : View(context, at
     }
 
     private fun Int.scaleX() = toFloat() / maxX * (width - GRAPH_MARGIN)
-    private fun Int.scaleY() = toFloat() / maxY * (height - GRAPH_MARGIN)
+    private fun Float.scaleY() = this / maxY * (height - GRAPH_MARGIN)
 
     companion object {
         const val GRAPH_MARGIN = 6f
